[cleanup] refactor relationship lookup/cache code
This helps us remove state from the model classes themselves. This helps
align us towards ember's vision of how they want classes to be
injectable in the future.
