+++ tests/integration/serializers/json-serializer-test.js
@@ -26,9 +26,9 @@ module("integration/serializer/json - JSONSerializer", {
       comment:  Comment,
       favorite: Favorite
     });
-    env.store.modelFor('post');
-    env.store.modelFor('comment');
-    env.store.modelFor('favorite');
+    Post = env.store.modelFor('post');
+    Comment = env.store.modelFor('comment');
+    Favorite = env.store.modelFor('favorite');
   },
 
   afterEach() {
@@ -297,7 +297,7 @@ test('Serializer should respect the attrs hash when extracting records', functio
 });
 
 test('Serializer should map `attrs` attributes directly when keyForAttribute also has a transform', function(assert) {
-  Post = DS.Model.extend({
+  Post.reopen({
     authorName: DS.attr('string')
   });
   env = setupStore({
